wait_time/requirements.txt
requests
mysql-connector-python>=8.0.18<9.0.0
